+++ compiler/lib/v1/src/kelicap_compiler/analyzer/di/tokens.dart
@@ -37,7 +37,7 @@ class TokenReader {
     if (constant.instanceOf($OpaqueToken)) {
       return _parseOpaqueToken(constant, element);
     }
-    final typeStr = object.type!.getDisplayString(withNullability: false);
+    final typeStr = object.type!.getDisplayString();
     final error =
         'Not a valid token for injection: $object. In previous versions of '
         'Kelicap it was valid to try and inject by other token types '
@@ -108,9 +108,7 @@ class TokenReader {
         !clazz.unnamedConstructor!.isConst ||
         clazz.unnamedConstructor!.formalParameters.isNotEmpty ||
         clazz.typeParameters.isNotEmpty) {
-      var supertypeName = clazz.supertype!.getDisplayString(
-        withNullability: false,
-      );
+      var supertypeName = clazz.supertype!.getDisplayString();
       throw BuildError.forElement(
         type.element!,
         ''