+++ kelicap/lib/src/core/exception_handler.dart
@@ -1,6 +1,5 @@
 import 'dart:js_interop';
 
-import 'package:web/helpers.dart';
 import 'package:web/web.dart';
 
 /// Provides a hook for receiving unhandled errors/exceptions.