+++ compiler/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 1.0.1
+
+- fix: Replaced deprecated methods used to get type string in di module
+
 ## 1.0.0
 
 - Initial release