+++ kelicap/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 1.0.1
+
+- Remove usage of querySelector and directly use document.querySelector
+- Update README
+
 ## 1.0.0
 
 - Initial release